hack: Fix the installation script
This was problematic as the `${NAMESPACE}` environment variable was being referencing throughout the script, but the `$namespace` was the initialization variable. Updated the initialization variable fixed that, and added creating the Prometheus recording rules to the list of resources that needs to be created. Note: due to validation logic in the openshift monitoring stack, this PrometheusRule custom resource, which holds the list of metering recording rules, needs to be stored in the openshift-monitoring namespace or else those rules won't be respected or available in the `/rules` API endpoint.