go.*: Add the client-go transporter package
The default http.RoundTripper implementation from the net/http package does not support
passing a bearer token file when instantiating a transport configuration. In the case
with an OCP cluster, the monitoring stack has TLS enabled by default and bearer token
authentication is the easiest way to hit the various Servies/Routes/etc. for Prometheus
metrics.