hack: Fix the installation script for clusterrole* resources
Update the hack/install.sh bash script to properly handle the clusterrole and clusterrolebinding resources correctly.

Before, the conditional wrapping the oc create clause was always evaluating to true despite no clusterrole or clusterrolebinding being present with the `app=runner` label.

After looking into this a bit, it appears that doing any sort of label selecting operation will always evaluate to false, even when a resource exists with that label.

In order to correctly evaluate, attempt to find the label we're interested in by piping to grep/awk and then re-getting that resource by name, instead of by label.