cmd/driver: Fix issue where the batch insert operation was hanging
Update the populatePostgresTables function logic, fixing an issue where the batch insert operation was hanging the process from completion.

This had been problematic for a while, and my first assumption was that the batch queue size was getting to large, which resulted in deadlock.

The actual problem was that we were spawning several connections to make that batch insert call, but we weren't closing that connection after
successfully inserting, which is pretty obvious in hindsight. Adding a defer call that closes that connection speeds up the program quite a bit
and fixes the problem where the connection was hanging.

Also, while debugging, added some artifacts for timing the batch insert call but there's probably some value in keeping those.