hack,manifests: Migrate to running the binary as a CronJob instead of a Deployment
Updates the hack/install.sh bash script to create a CronJob that's responsible for spawning a Job that runs a specific container every five minutes.

The cmd/driver implementation only runs to completion at the moment instead of being wrapped in a go routine that has time-based logic built into it.

As a workaround in the meantime, just spawn off several Jobs that are responsible for running the postgres runner container every five minutes to get a sense of how much data can be populated in Postgres in the span of ~14 hours (or whenever the cluster reaches EOL).